V1.3.1 beta

+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/services/impl/PublicStepsServiceImpl.java
@@ -16,6 +16,7 @@ import org.cloud.sonic.controller.services.impl.base.SonicServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -134,9 +135,14 @@ public class PublicStepsServiceImpl extends SonicServiceImpl<PublicStepsMapper, 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public boolean deleteByProjectId(int projectId) {
-        PublicSteps publicSteps = lambdaQuery().eq(PublicSteps::getProjectId, projectId).one();
-        publicStepsStepsMapper.delete(new LambdaQueryWrapper<PublicStepsSteps>()
-                .eq(PublicStepsSteps::getPublicStepsId, publicSteps.getId()));
-        return delete(publicSteps.getId());
+        List<PublicSteps> publicSteps = lambdaQuery().eq(PublicSteps::getProjectId, projectId).list();
+        for (PublicSteps publicStep : publicSteps) {
+            if (!ObjectUtils.isEmpty(publicSteps)) {
+                publicStepsStepsMapper.delete(new LambdaQueryWrapper<PublicStepsSteps>()
+                        .eq(PublicStepsSteps::getPublicStepsId, publicStep.getId()));
+                delete(publicStep.getId());
+            }
+        }
+        return true;
     }
 }
